test: share client fixtures
